Move aproxy test to own module
This is to avoid interference with other tests in the same module.
